EES-4482 Add caching to new redirects endpoint
