+++ src/GovUk.Education.ExploreEducationStatistics.Content.Api/Controllers/RedirectsController.cs
@@ -1,7 +1,7 @@
 #nullable enable
 using System.Threading.Tasks;
 using GovUk.Education.ExploreEducationStatistics.Common.Extensions;
-using GovUk.Education.ExploreEducationStatistics.Content.Services.Interfaces;
+using GovUk.Education.ExploreEducationStatistics.Content.Services.Interfaces.Cache;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels;
 using Microsoft.AspNetCore.Mvc;
 
@@ -10,18 +10,18 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Api.Controllers
     [Route("api")]
     public class RedirectsController : ControllerBase
     {
-        private readonly IRedirectsService _redirectsService;
+        private readonly IRedirectsCacheService _redirectsCacheService;
 
         public RedirectsController(
-            IRedirectsService redirectsService)
+            IRedirectsCacheService redirectsCacheService)
         {
-            _redirectsService = redirectsService;
+            _redirectsCacheService = redirectsCacheService;
         }
 
         [HttpGet("redirects")]
-        public async Task<ActionResult<RedirectsViewModel>> ListRedirects()
+        public async Task<ActionResult<RedirectsViewModel>> List()
         {
-            return await _redirectsService.List()
+            return await _redirectsCacheService.List()
                 .HandleFailuresOrOk();
         }
     }
