+++ src/GovUk.Education.ExploreEducationStatistics.Content.Model/Repository/Interfaces/IMethodologyVersionRepository.cs
@@ -22,7 +22,5 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Model.Repository.In
         Task<bool> IsLatestPublishedVersion(MethodologyVersion methodologyVersion);
 
         Task<bool> IsToBePublished(MethodologyVersion methodologyVersion);
-
-        Task PublicationTitleOrSlugChanged(Guid publicationId, string originalSlug, string updatedTitle, string updatedSlug);
     }
 }
