+++ src/GovUk.Education.ExploreEducationStatistics.Content.Services.Tests/MethodologyServiceTests.cs
@@ -10,8 +10,10 @@ using GovUk.Education.ExploreEducationStatistics.Common.Utils;
 using GovUk.Education.ExploreEducationStatistics.Content.Model;
 using GovUk.Education.ExploreEducationStatistics.Content.Model.Database;
 using GovUk.Education.ExploreEducationStatistics.Content.Model.Repository.Interfaces;
+using GovUk.Education.ExploreEducationStatistics.Content.Services.Interfaces.Cache;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.Mappings;
 using GovUk.Education.ExploreEducationStatistics.Content.Services.ViewModels;
+using Microsoft.EntityFrameworkCore;
 using Moq;
 using Xunit;
 using static GovUk.Education.ExploreEducationStatistics.Common.Services.CollectionUtils;
@@ -677,17 +679,487 @@ namespace GovUk.Education.ExploreEducationStatistics.Content.Services.Tests
             VerifyAllMocks(methodologyVersionRepository);
         }
 
+        [Fact]
+        public async Task PublicationTitleOrSlugChanged()
+        {
+            var publicationId = Guid.NewGuid();
+            var originalVersionId = Guid.NewGuid();
+            var latestPublishedVersionId = Guid.NewGuid();
+
+            var contentDbContextId = Guid.NewGuid().ToString();
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var publicationMethodology = new PublicationMethodology
+                {
+                    PublicationId = publicationId,
+                    Owner = true,
+                    Methodology = new Methodology
+                    {
+                        LatestPublishedVersionId = latestPublishedVersionId,
+                        OwningPublicationTitle = "Original Title",
+                        OwningPublicationSlug = "original-slug",
+                        Versions = new List<MethodologyVersion>
+                        {
+                            new()
+                            {
+                                Id = originalVersionId,
+                                Version = 0,
+                            },
+                            new()
+                            {
+                                Id = latestPublishedVersionId,
+                                Version = 1,
+                                PreviousVersionId = originalVersionId,
+                            },
+                        },
+                    },
+                };
+
+                await contentDbContext.PublicationMethodologies.AddAsync(publicationMethodology);
+                await contentDbContext.SaveChangesAsync();
+            }
+
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var redirectsCacheService = new Mock<IRedirectsCacheService>(MockBehavior.Strict);
+                redirectsCacheService.Setup(mock => mock.UpdateRedirects())
+                    .ReturnsAsync(new RedirectsViewModel(new List<MethodologyRedirectViewModel>()));
+
+                var service = SetupMethodologyService(contentDbContext,
+                    redirectsCacheService: redirectsCacheService.Object);
+                await service.PublicationTitleOrSlugChanged(publicationId, "original-slug", "New Title", "new-slug");
+            }
+
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var publicationMethodology = await contentDbContext
+                    .PublicationMethodologies
+                    .Include(m => m.Methodology.Versions)
+                    .SingleAsync(m => m.PublicationId == publicationId);
+
+                Assert.Equal("New Title", publicationMethodology.Methodology.OwningPublicationTitle);
+                Assert.Equal("new-slug", publicationMethodology.Methodology.OwningPublicationSlug);
+
+                // As no AlternativeTitle or AlternativeSlug set, the MethodologyVersion's title and slug will also change
+                Assert.Equal("New Title", publicationMethodology.Methodology.Versions[1].Title);
+                Assert.Equal("new-slug", publicationMethodology.Methodology.Versions[1].Slug);
+
+                // As methodology is published and it's slug has changed, a redirect is created for LatestPublishedVersion
+                var methodologyRedirects = await contentDbContext.MethodologyRedirects
+                    .ToListAsync();
+                var methodologyRedirect = Assert.Single(methodologyRedirects);
+                Assert.Equal(latestPublishedVersionId, methodologyRedirect.MethodologyVersionId);
+                Assert.Equal("original-slug", methodologyRedirect.Slug);
+            }
+        }
+
+        [Fact]
+        public async Task PublicationTitleOrSlugChanged_NoMethodologyRedirectAsMethodologyUnpublished()
+        {
+            var publicationId = Guid.NewGuid();
+
+            var contentDbContextId = Guid.NewGuid().ToString();
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var publicationMethodology = new PublicationMethodology
+                {
+                    PublicationId = publicationId,
+                    Owner = true,
+                    Methodology = new Methodology
+                    {
+                        LatestPublishedVersionId = null,
+                        OwningPublicationTitle = "Original Title",
+                        OwningPublicationSlug = "original-slug",
+                        Versions = ListOf(new MethodologyVersion()),
+                    },
+                };
+
+                await contentDbContext.PublicationMethodologies.AddAsync(publicationMethodology);
+                await contentDbContext.SaveChangesAsync();
+            }
+
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var redirectsCacheService = new Mock<IRedirectsCacheService>(MockBehavior.Strict);
+                redirectsCacheService.Setup(mock => mock.UpdateRedirects())
+                    .ReturnsAsync(new RedirectsViewModel(new List<MethodologyRedirectViewModel>()));
+
+                var service = SetupMethodologyService(contentDbContext,
+                    redirectsCacheService: redirectsCacheService.Object);
+                await service.PublicationTitleOrSlugChanged(publicationId,
+                    "original-slug", "New Title", "new-slug");
+            }
+
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var publicationMethodology = await contentDbContext
+                    .PublicationMethodologies
+                    .Include(m => m.Methodology.Versions)
+                    .SingleAsync(m => m.PublicationId == publicationId);
+
+                Assert.Equal("New Title", publicationMethodology.Methodology.OwningPublicationTitle);
+                Assert.Equal("new-slug", publicationMethodology.Methodology.OwningPublicationSlug);
+
+                // As the Publication's Title and Slug changed, and no AlternateTitle/Slug set,
+                // the methodology's title and slug will also change
+                Assert.Equal("New Title", publicationMethodology.Methodology.Versions[0].Title);
+                Assert.Equal("new-slug", publicationMethodology.Methodology.Versions[0].Slug);
+
+                // Methodology is unpublished, so no redirect
+                var methodologyRedirects = await contentDbContext.MethodologyRedirects
+                    .ToListAsync();
+                Assert.Empty(methodologyRedirects);
+            }
+        }
+
+
+
+        [Fact]
+        public async Task PublicationTitleOrSlugChanged_DoesNotAffectUnrelatedMethodologies()
+        {
+            var publicationId = Guid.NewGuid();
+            var unrelatedPublicationId = Guid.NewGuid();
+
+            var contentDbContextId = Guid.NewGuid().ToString();
+
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var publicationMethodology = new PublicationMethodology
+                {
+                    PublicationId = publicationId,
+                    Owner = true,
+                    Methodology = new Methodology
+                    {
+                        Versions = ListOf(new MethodologyVersion
+                        {
+                            Status = Draft
+                        }),
+                        OwningPublicationTitle = "Original Title",
+                        OwningPublicationSlug = "original-slug",
+                    }
+                };
+
+                var unrelatedPublicationMethodology = new PublicationMethodology
+                {
+                    PublicationId = unrelatedPublicationId,
+                    Owner = true,
+                    Methodology = new Methodology
+                    {
+                        Versions = ListOf(new MethodologyVersion
+                        {
+                            Status = Draft
+                        }),
+                        OwningPublicationTitle = "Original Title",
+                        OwningPublicationSlug = "original-slug",
+                    }
+                };
+
+                await contentDbContext.PublicationMethodologies.AddRangeAsync(
+                    publicationMethodology,
+                    unrelatedPublicationMethodology);
+                await contentDbContext.SaveChangesAsync();
+            }
+
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var redirectsCacheService = new Mock<IRedirectsCacheService>(MockBehavior.Strict);
+                redirectsCacheService.Setup(mock => mock.UpdateRedirects())
+                    .ReturnsAsync(new RedirectsViewModel(new List<MethodologyRedirectViewModel>()));
+
+                var service = SetupMethodologyService(contentDbContext,
+                    redirectsCacheService: redirectsCacheService.Object);
+                await service.PublicationTitleOrSlugChanged(publicationId, "original-slug", "New Title", "new-slug");
+            }
+
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var publicationMethodology = await contentDbContext
+                    .PublicationMethodologies
+                    .Include(m => m.Methodology)
+                    .SingleAsync(m => m.PublicationId == unrelatedPublicationId);
+
+                // This Methodology was not related to the Publication being updated, and so was not affected by the update.
+                Assert.Equal("Original Title", publicationMethodology.Methodology.OwningPublicationTitle);
+                Assert.Equal("original-slug", publicationMethodology.Methodology.OwningPublicationSlug);
+            }
+        }
+
+        [Fact]
+        public async Task PublicationTitleOrSlugChanged_DoesNotAffectUnownedMethodologies()
+        {
+            var publicationId = Guid.NewGuid();
+
+            var contentDbContextId = Guid.NewGuid().ToString();
+
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var publicationMethodology = new PublicationMethodology
+                {
+                    PublicationId = publicationId,
+                    Owner = false,
+                    Methodology = new Methodology
+                    {
+                        Versions = ListOf(new MethodologyVersion()),
+                        OwningPublicationTitle = "Original Title",
+                        OwningPublicationSlug = "original-slug",
+                    }
+                };
+
+                await contentDbContext.PublicationMethodologies.AddAsync(publicationMethodology);
+                await contentDbContext.SaveChangesAsync();
+            }
+
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var service = SetupMethodologyService(contentDbContext);
+                await service.PublicationTitleOrSlugChanged(publicationId,
+                    "original-slug", "New Title", "new-slug");
+            }
+
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var publicationMethodology = await contentDbContext
+                    .PublicationMethodologies
+                    .Include(m => m.Methodology)
+                    .SingleAsync(m => m.PublicationId == publicationId);
+
+                // This Methodology was not owned by the Publication being updated, and so was not affected by the update.
+                Assert.Equal("Original Title", publicationMethodology.Methodology.OwningPublicationTitle);
+                Assert.Equal("original-slug", publicationMethodology.Methodology.OwningPublicationSlug);
+            }
+        }
+
+        [Fact]
+        public async Task PublicationTitleOrSlugChanged_MethodologySlugHasAlreadyBeenAmended()
+        {
+            var publicationId = Guid.NewGuid();
+            var latestPublishedVersionId = Guid.NewGuid();
+
+            var contentDbContextId = Guid.NewGuid().ToString();
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var publicationMethodology = new PublicationMethodology
+                {
+                    PublicationId = publicationId,
+                    Owner = true,
+                    Methodology = new Methodology
+                    {
+                        LatestPublishedVersionId = latestPublishedVersionId,
+                        OwningPublicationTitle = "Original title",
+                        OwningPublicationSlug = "original-slug",
+                        Versions = ListOf(new MethodologyVersion
+                        {
+                            Id = latestPublishedVersionId,
+                            AlternativeSlug = "alternative-slug",
+                        }),
+                    }
+                };
+
+                await contentDbContext.PublicationMethodologies.AddAsync(publicationMethodology);
+                await contentDbContext.SaveChangesAsync();
+            }
+
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var redirectsCacheService = new Mock<IRedirectsCacheService>(MockBehavior.Strict);
+                redirectsCacheService.Setup(mock => mock.UpdateRedirects())
+                    .ReturnsAsync(new RedirectsViewModel(new List<MethodologyRedirectViewModel>()));
+
+                var service = SetupMethodologyService(contentDbContext,
+                    redirectsCacheService: redirectsCacheService.Object);
+                await service.PublicationTitleOrSlugChanged(publicationId, "original-slug", "New Title", "new-slug");
+            }
+
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var publicationMethodology = await contentDbContext
+                    .PublicationMethodologies
+                    .Include(m => m.Methodology)
+                    .ThenInclude(m => m.Versions)
+                    .SingleAsync(m => m.PublicationId == publicationId);
+
+                Assert.Equal("New Title", publicationMethodology.Methodology.OwningPublicationTitle);
+                Assert.Equal("new-slug", publicationMethodology.Methodology.OwningPublicationSlug);
+
+                // The MethodologyVersion has already had an AlternativeSlug set. It doesn't have a AlternativeTitle
+                // set. So the MethodologyVersion title is updated, but the slug remains the same.
+                Assert.Equal("New Title", publicationMethodology.Methodology.Versions[0].Title);
+                Assert.Equal("alternative-slug", publicationMethodology.Methodology.Versions[0].Slug);
+
+                // No redirect created as slug hasn't changed
+                var methodologyRedirects = await contentDbContext.MethodologyRedirects
+                    .ToListAsync();
+                Assert.Empty(methodologyRedirects);
+            }
+        }
+
+        [Fact]
+        public async Task PublicationTitleOrSlugChanged_MethodologyAlreadyPubliclyAvailable()
+        {
+            var publicationId = Guid.NewGuid();
+            var latestPublishedVersionId = Guid.NewGuid();
+
+            var contentDbContextId = Guid.NewGuid().ToString();
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var publicationMethodology = new PublicationMethodology
+                {
+                    Publication = new Publication
+                    {
+                        Id = publicationId,
+                        LatestPublishedRelease = new Release()
+                    },
+                    Owner = true,
+                    Methodology = new Methodology
+                    {
+                        LatestPublishedVersionId = latestPublishedVersionId,
+                        Versions = ListOf(new MethodologyVersion
+                        {
+                            Id = latestPublishedVersionId,
+                        }),
+                        OwningPublicationTitle = "Original title",
+                        OwningPublicationSlug = "original-slug",
+                    }
+                };
+
+                await contentDbContext.PublicationMethodologies.AddAsync(publicationMethodology);
+                await contentDbContext.SaveChangesAsync();
+            }
+
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var redirectsCacheService = new Mock<IRedirectsCacheService>(MockBehavior.Strict);
+                redirectsCacheService.Setup(mock => mock.UpdateRedirects())
+                    .ReturnsAsync(new RedirectsViewModel(new List<MethodologyRedirectViewModel>()));
+
+                var service = SetupMethodologyService(contentDbContext,
+                    redirectsCacheService: redirectsCacheService.Object);
+                await service.PublicationTitleOrSlugChanged(publicationId, "original-slug", "New Title", "new-slug");
+            }
+
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var publicationMethodology = await contentDbContext
+                    .PublicationMethodologies
+                    .Include(m => m.Methodology)
+                    .SingleAsync(m => m.PublicationId == publicationId);
+
+                Assert.Equal("New Title", publicationMethodology.Methodology.OwningPublicationTitle);
+                Assert.Equal("new-slug", publicationMethodology.Methodology.OwningPublicationSlug);
+
+                var redirect = await contentDbContext
+                    .MethodologyRedirects
+                    .SingleAsync();
+
+                Assert.Equal(latestPublishedVersionId, redirect.MethodologyVersionId);
+                Assert.Equal("original-slug", redirect.Slug);
+            }
+        }
+
+        [Fact]
+        public async Task PublicationTitleOrSlugChanged_NewMethodologyRedirectSlugMatchesExistingRedirectSlug()
+        {
+            var publicationId = Guid.NewGuid();
+            var latestPublishedVersionId = Guid.NewGuid();
+            var amendmentVersionId = Guid.NewGuid();
+
+            var contentDbContextId = Guid.NewGuid().ToString();
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var publicationMethodology = new PublicationMethodology
+                {
+                    Publication = new Publication
+                    {
+                        Id = publicationId,
+                        LatestPublishedRelease = new Release()
+                    },
+                    Owner = true,
+                    Methodology = new Methodology
+                    {
+                        LatestPublishedVersionId = latestPublishedVersionId,
+                        Versions = new List<MethodologyVersion>
+                        {
+                            new ()
+                            {
+                                Id = latestPublishedVersionId,
+                                Version = 0,
+                            },
+                            new ()
+                            {
+                                Id = amendmentVersionId,
+                                AlternativeSlug = "newer-slug",
+                                Version = 1,
+                                PreviousVersionId = latestPublishedVersionId,
+                            },
+                        },
+                        OwningPublicationTitle = "Original title",
+                        OwningPublicationSlug = "original-slug",
+                    }
+                };
+
+                var methodologyRedirect = new MethodologyRedirect
+                {
+                    MethodologyVersion = publicationMethodology.Methodology.Versions[1],
+                    Slug = "original-slug",
+                };
+
+                await contentDbContext.PublicationMethodologies.AddAsync(publicationMethodology);
+                await contentDbContext.MethodologyRedirects.AddAsync(methodologyRedirect);
+                await contentDbContext.SaveChangesAsync();
+            }
+
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var redirectsCacheService = new Mock<IRedirectsCacheService>(MockBehavior.Strict);
+                redirectsCacheService.Setup(mock => mock.UpdateRedirects())
+                    .ReturnsAsync(new RedirectsViewModel(new List<MethodologyRedirectViewModel>()));
+
+                var service = SetupMethodologyService(contentDbContext,
+                    redirectsCacheService: redirectsCacheService.Object);
+                await service.PublicationTitleOrSlugChanged(publicationId, "original-slug",
+                    "New Title", "new-slug");
+            }
+
+            await using (var contentDbContext = InMemoryContentDbContext(contentDbContextId))
+            {
+                var publicationMethodology = await contentDbContext
+                    .PublicationMethodologies
+                    .Include(m => m.Methodology.Versions)
+                    .SingleAsync(m => m.PublicationId == publicationId);
+
+                Assert.Equal("New Title", publicationMethodology.Methodology.OwningPublicationTitle);
+                Assert.Equal("new-slug", publicationMethodology.Methodology.OwningPublicationSlug);
+
+                var redirects = await contentDbContext
+                    .MethodologyRedirects
+                    .ToListAsync();
+
+                // We remove the redundant redirect from the unpublished amendment. This is necessary because
+                // if the unpublished amendment's slug is changed again from "newer-slug" to "even-newer-slug", no
+                // new redirect will be created, as the amendment already has a redirect from "original-slug". This
+                // would mean that when the amendment is published, and the methodology's slug changes to
+                // "even-newer-slug", there will be no redirect from "new-slug".
+                var redirect = Assert.Single(redirects);
+
+                Assert.Equal(latestPublishedVersionId, redirect.MethodologyVersionId);
+                Assert.Equal("original-slug", redirect.Slug);
+            }
+        }
+
+
         private static MethodologyService SetupMethodologyService(
             ContentDbContext contentDbContext,
             IPersistenceHelper<ContentDbContext>? contentPersistenceHelper = null,
             IMethodologyVersionRepository? methodologyVersionRepository = null,
-            IMapper? mapper = null)
+            IMapper? mapper = null,
+            IRedirectsCacheService? redirectsCacheService = null)
         {
             return new(
                 contentDbContext,
                 contentPersistenceHelper ?? new PersistenceHelper<ContentDbContext>(contentDbContext),
                 mapper ?? MapperUtils.MapperForProfile<MappingProfiles>(),
-                methodologyVersionRepository ?? Mock.Of<IMethodologyVersionRepository>(MockBehavior.Strict)
+                methodologyVersionRepository ?? Mock.Of<IMethodologyVersionRepository>(MockBehavior.Strict),
+                redirectsCacheService ?? Mock.Of<IRedirectsCacheService>(MockBehavior.Strict)
             );
         }
     }
